requirements.txt
edge_tts
pygame
pyaudio
wave
keyboard
groq
python-dotenv
ollama
transformers
torch
